# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Ambulanzsystem V1 API
  version: 0.1.0
  description: |
    Frozen Stage-0 contract for the Ambulanzsystem V1 (see docs/plan/master-plan.md).
    Locked decisions D1-D8 apply. JSON property names are serialized exactly as
    written here (no automatic camelCase transformation server-side).

    Auth token types (JWT `type` claim): `admin`, `leitstelle`, `user`, `qr`.
    Policies referenced below:
    - AdminOnly: type=admin
    - LeitstelleOrAdmin: type in (admin, leitstelle)
    - AuthenticatedUser: type in (admin, leitstelle, user)
    - TriageWrite: type in (admin, leitstelle, user, qr)

    Canonical triage values: rot | gelb | gruen | schwarz (ASCII, D5).
    `blau` and Unicode `grün` are rejected at the write boundary.
servers:
  - url: http://localhost:5000
    description: local backend
  - url: http://localhost:4010
    description: Prism mock server
security:
  - bearerAuth: []
...
paths:
...
  /api/persons/{patientId}/ambulanzprotokoll-page1:
    get:
      tags: [ambulanzprotokoll]
      summary: Get the page-1 record, or a default-shaped empty draft if none saved yet. Read-audit-logged. Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/patientIdPath" }]
      responses:
        "200":
          description: Form record (formState conforms to schemas/ambulanzprotokoll-page1.schema.json).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProtokollRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Patient does not exist. }
    put:
      tags: [ambulanzprotokoll]
      summary: >
        Upsert the page-1 record. Draft accepts any partial data; missing branches
        are merged with the default state; concurrent saves merge field-level LWW
        (older non-empty values survive empty newer fields, NFR-SAFE-09).
        status=finalized stamps finalizedAt and returns soft-validation warnings —
        never a hard block for empty fields. Post-finalize saves stay allowed
        (Leitstelle/Admin any form; responder their own records); all audit-logged.
        Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/patientIdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, formState]
              properties:
                status: { type: string, enum: [draft, finalized] }
                formState:
                  type: object
                  description: Must conform to schemas/ambulanzprotokoll-page1.schema.json.
                clientUpdatedAt: { type: string, format: date-time, description: Local edit timestamp for newer-wins comparison. }
      responses:
        "200":
          description: Saved record (merged server state) plus any finalize warnings.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ProtokollRecord"
                  - type: object
                    properties:
                      warnings:
                        type: array
                        items: { type: string }
        "400": { description: Status outside draft|finalized, or structurally unstorable payload. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Patient does not exist. }
  /api/persons/{patientId}/ambulanzprotokoll-page1/export:
    get:
      tags: [ambulanzprotokoll]
      summary: >
        JSON export (V1 export format, FR-DOC-09). Wraps formState with
        patient/event metadata, generation timestamp, and an actor watermark
        (NFR-SEC-10). The export is persisted server-side (long-term archive,
        FR-DOC-14) and the action is audit-logged. Policy TriageWrite; responders
        may export patients they recorded.
      parameters: [{ $ref: "#/components/parameters/patientIdPath" }]
      responses:
        "200":
          description: Export document.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProtokollExport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------------------------------- body parts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    patientIdPath:
      name: patientId
      in: path
      required: true
      schema: { type: integer }
  responses:
    Unauthorized:
      description: Missing/invalid/expired/revoked credentials. Expired access blocks new actions and reads (FR-AUTH-09).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [status, message]
      properties:
        status: { type: string, const: error }
        message: { type: string }
...
    ProtokollRecord:
      type: object
      required: [patientId, status, formState, updatedAt]
      properties:
        patientId: { type: integer }
        status: { type: string, enum: [draft, finalized] }
        formState:
          type: object
          description: Conforms to schemas/ambulanzprotokoll-page1.schema.json.
        updatedAt: { type: string, format: date-time }
        finalizedAt: { type: [string, "null"], format: date-time }
    ProtokollExport:
      type: object
      required: [metadata, protokoll]
      properties:
        metadata:
          type: object
          required: [patientId, operationSceneId, generatedAt, watermark]
          properties:
            patientId: { type: integer }
            humanReadableId: { type: [string, "null"] }
            operationSceneId: { type: integer }
            operationSceneName: { type: string }
            eventSceneId: { type: integer }
            generatedAt: { type: string, format: date-time }
            watermark: { type: string, description: Identifies the exporting actor (NFR-SEC-10). }
            schemaVersion: { type: string }
        protokoll: { $ref: "#/components/schemas/ProtokollRecord" }
...